Implement JWT authentication in authenticateToken middleware
Added a middleware called authenticateToken to handle JWT authentication in protected routes. The middleware checks if a JWT token is provided in the request's Authorization header. If no token is provided or if the token is invalid, it returns a status code of 401 (Unauthorized) or 403 (Forbidden) respectively. jsonwebtoken library is used to verify the validity of the JWT token using the secret key stored in the SECRET_JWT environment variable.