Implement authentication middleware for JWT tokens
This commit introduces a new middleware function named uthenticateJWT responsible for handling authentication using JSON Web Tokens (JWT). The middleware is designed to verify the presence and validity of both access and refresh tokens in the request cookies. If both tokens are absent, the middleware returns a 401 Unauthorized response. When only the access token is provided, it is validated using the SECRET_JWT and the user object is set in the request. In case only the refresh token is found, it undergoes validation with the REFRESH_JWT, following which new access and refresh tokens are generated and set in the response cookies. Any encountered errors during token verification prompt the middleware to clear the cookies and respond with a 401 Unauthorized status code.

This implementation aims to bolster the security of the application by ensuring robust authentication mechanisms and seamless token management.